add extra security to delete course

+++ app/controllers/courses_controller.rb
@@ -82,6 +82,9 @@ class CoursesController < ApplicationController
   end
 
   def delete
+    return redirect_to courses_path, alert: 'You do not have access to this page.' unless @role == 'instructor'
+    return redirect_to courses_path, alert: 'Extensions are not enabled for this course.' unless @course.course_settomgs&.enable_extensions
+
     assignments = Assignment.where(course_to_lms_id: CourseToLms.where(course_id: @course.id).select(:id))
     Extension.where(assignment_id: assignments.select(:id)).destroy_all
     assignments.destroy_all
